create separate Docker struct for each package using it
