+++ bridge/docker.go
@@ -1,0 +1,5 @@
+package bridge
+
+type Docker struct {
+	Image string `json:"image" yaml:"image"`
+}
