+++ flannel/docker.go
@@ -1,4 +1,4 @@
-package docker
+package flannel
 
 type Docker struct {
 	Image string `json:"image" yaml:"image"`
